Move ShellOption to its own module and implement UI loop
