+++ src/user_interface.rs
@@ -1,1 +1,2 @@
 pub mod command_line;
+pub mod shell_option;
