+++ src/user_interface/command_line.rs
@@ -1,128 +1,105 @@
-use std::str::Split;
+use std::io::{self, Write};
 
-enum ShellOption {
-    /// Initial option state. Should only be used for the startup of the main loop to prevent anything from happening
-    Initial,
-    Help,
-    Exit,
-    Create,
-    Read(ReadType),
-    /// Has update type and an id
-    Update(UpdateType, i32),
-    /// Delete `id`
-    Delete(i32),
-}
+use mysql::prelude::Queryable;
+use time::{Date, Duration};
 
-enum ReadType {
-    All,
-    Pending,
-    /// Search for specific string
-    Search(String),
-    /// Show only `id`
-    One(i32),
-}
+use crate::repository::job_application_repository::{
+    insert_job_application, HumanResponse, JobApplication,
+};
 
-enum UpdateType {
-    HumanResponse,
-    Other,
-}
+use super::shell_option::{ReadType, ShellOption, UpdateType};
 
-impl TryFrom<&str> for ShellOption {
-    type Error = String;
+/// The main loop that runs the prompt
+/// Will exit if there is an  
+pub fn main_loop<C: Queryable>(conn: &mut C) -> Result<(), std::io::Error> {
+    let stdin = std::io::stdin();
 
-    fn try_from(value: &str) -> Result<Self, Self::Error> {
-        let mut value_words = value.split(' ');
-        if let Some(command_word) = value_words.next() {
-            match command_word {
-                "help" | "h" => Ok(Self::Help),
-                "exit" | "quit" => Ok(Self::Exit),
-                "create" => Ok(Self::Create),
-                // For read command, parse the read type as well
-                "read" => Ok(Self::Read(ReadType::try_from(&mut value_words)?)),
-                // For update command, parse the read type and the id
-                "update" => {
-                    let update_type = UpdateType::try_from(&mut value_words)?;
-                    if let Some(id_str) = value_words.next() {
-                        match str::parse::<i32>(id_str) {
-                            Ok(id) => Ok(Self::Update(update_type, id)),
-                            Err(err_message) => Err(format!(
-                                "Unable to parse id '{id_str}'. Error: {err_message}"
-                            )),
-                        }
-                    } else {
-                        Err("ID is required for update".to_owned())
-                    }
-                }
-                // For delete, parse the id
-                "delete" => {
-                    if let Some(id_str) = value_words.next() {
-                        match str::parse::<i32>(id_str) {
-                            Ok(id) => Ok(Self::Delete(id)),
-                            Err(err_message) => Err(format!(
-                                "Unable to parse id '{id_str}'. Error: {err_message}"
-                            )),
-                        }
-                    } else {
-                        Err("ID is required for delete".to_owned())
-                    }
-                }
-                _ => Err("Invalid command".to_owned()),
-            }
-        } else {
-            Err("No command given".to_owned())
-        }
+    print!("ats tracking> ");
+    io::stdout().flush().unwrap();
+    for line in stdin.lines() {
+        let input = line?;
+
+        match ShellOption::try_from(input.as_str()) {
+            // If we should exit, do that
+            Ok(ShellOption::Exit) => break,
+            // If there is an error, print the error
+            Err(s) => println!("{}", s),
+            // For any other commands
+            Ok(command) => match command {
+                ShellOption::Help => help(),
+                ShellOption::Create => create(conn),
+                ShellOption::Read(read_type) => read(conn, read_type),
+                ShellOption::Update(update_type, id) => update(conn, update_type, id),
+                ShellOption::Delete(id) => delete(conn, id),
+                ShellOption::Exit => unreachable!(),
+            },
+        };
+
+        print!("ats tracking> ");
+        io::stdout().flush().unwrap();
     }
+
+    Ok(())
 }
 
-impl TryFrom<&mut Split<'_, char>> for ReadType {
-    type Error = String;
+fn help() {
+    print!(
+        "ATS Tracking System
 
-    fn try_from(value: &mut Split<'_, char>) -> Result<Self, Self::Error> {
-        if let Some(read_type) = value.next() {
-            match read_type {
-                "all" => Ok(Self::All),
-                "pending" => Ok(Self::Pending),
-                "search" => {
-                    if let Some(search_query) = value.next() {
-                        Ok(Self::Search(search_query.to_owned()))
-                    } else {
-                        Err("Search query is required for search".to_owned())
-                    }
-                }
-                "one" => {
-                    if let Some(id_str) = value.next() {
-                        match str::parse::<i32>(id_str) {
-                            Ok(id) => Ok(Self::One(id)),
-                            Err(err_message) => Err(format!(
-                                "Unable to parse id '{id_str}'. Error: {err_message}"
-                            )),
-                        }
-                    } else {
-                        Err("ID is required for read one".to_owned())
-                    }
-                }
-                _ => Err("Invalid read type".to_owned()),
-            }
-        } else {
-            Err("No read type given".to_owned())
-        }
-    }
+Available commands:
+  help | h
+  exit | quit
+  create
+  read (all | pending | search <search_query> | one <id>)
+  update (response | other) <id>
+  delete <id>
+"
+    );
 }
 
-impl TryFrom<&mut Split<'_, char>> for UpdateType {
-    type Error = String;
+fn create<C: Queryable>(conn: &mut C) {
+    // Declare the variables here to make sure I define all of them
+    let id: i32;
+    let source: String;
+    let company: String;
+    let job_title: String;
+    let application_date: Date;
+    let time_investment: Option<Duration>;
+    let automated_response: bool;
+    let human_response: HumanResponse;
+    let human_response_date: Option<Date>;
+    let application_website: Option<String>;
+    let notes: Option<String>;
 
-    fn try_from(value: &mut Split<'_, char>) -> Result<Self, Self::Error> {
-        if let Some(update_type) = value.next() {
-            match update_type {
-                "response" => Ok(Self::HumanResponse),
-                "other" => Ok(Self::Other),
-                _ => Err("Invalid update type".to_owned()),
-            }
-        } else {
-            Err("No update type given".to_owned())
-        }
-    }
+    // Initialize the fields
+    todo!();
+
+    // Construct the new application.
+    let new_application = JobApplication {
+        id,
+        source,
+        company,
+        job_title,
+        application_date,
+        time_investment,
+        automated_response,
+        human_response,
+        human_response_date,
+        application_website,
+        notes,
+    };
+
+    insert_job_application(new_application);
 }
 
-fn main_loop() {}
+fn read<C: Queryable>(conn: &mut C, read_type: ReadType) {
+    todo!();
+}
+
+fn update<C: Queryable>(conn: &mut C, update_type: UpdateType, id: i32) {
+    todo!();
+}
+
+fn delete<C: Queryable>(conn: &mut C, id: i32) {
+    todo!();
+}
